requirements.txt
einops==0.4.1
matplotlib==3.3.4
multiprocess==0.70.12.2
open3d==0.13.0
opencv-python==4.5.3.56
Pillow==8.3.2
plyfile==0.7.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
trimesh==3.9.41